refactor: reorganise codebase and split large modules for better structure
