+++ src/domain/badges.rs
