+++ src/domain/mod.rs
@@ -1,0 +1,2 @@
+pub mod badges;
+pub mod models;
