+++ src/http/handlers/mod.rs
@@ -1,3 +1,4 @@
+pub mod file_page;
 pub mod files;
 pub mod heartbeat;
 pub mod presence;
