+++ src/integrations/discord/mod.rs
@@ -3,11 +3,11 @@ use std::{collections::BTreeMap, sync::Arc};
 use chrono::{DateTime, Utc};
 use serenity::{
     all::{
-        ButtonStyle, CommandDataOptionValue, CommandInteraction, CommandOptionType,
-        ComponentInteraction, CreateActionRow, CreateButton, CreateCommand, CreateCommandOption,
+        ButtonStyle, CommandDataOptionValue, CommandInteraction,
+        ComponentInteraction, CreateActionRow, CreateButton,
         CreateEmbed, CreateEmbedAuthor, CreateEmbedFooter, CreateInteractionResponse,
         CreateInteractionResponseMessage, GatewayIntents, Guild, GuildId, Interaction, Member,
-        OnlineStatus, Permissions, Presence, Ready, User as DiscordUser,
+        OnlineStatus, Presence, Ready, User as DiscordUser,
     },
     async_trait,
     client::{Client, Context, EventHandler},
@@ -17,10 +17,24 @@ use serenity::{
 use sqlx::{Row, query, query_as};
 
 use crate::{
-    app_state::AppState,
-    error::AppError,
-    models::User,
+    app::state::AppState,
+    app::error::AppError,
+    domain::models::User,
     utils::ids::{generate_api_token, is_valid_hex_colour, normalise_url_mode},
+};
+
+mod commands;
+mod helpers;
+
+use commands::{
+    admin_files_command, admin_users_command, blacklist_command, config_command,
+    delete_file_command, files_command, kv_command, ping_command, profile_command,
+    register_command, register_success_message, token_command,
+};
+use helpers::{
+    admin_files_pagination_row, chunk_buttons, clamp_page, is_admin,
+    owned_pagination_components, parse_owned_page_custom_id, respond, respond_component,
+    respond_component_embed, respond_embed, total_pages, update_component,
 };
 
 struct ShardManagerContainer;
@@ -1411,406 +1425,3 @@ fn validate_kv_value(value: &str) -> Result<(), AppError> {
     Ok(())
 }
 
-fn is_admin(permissions: Option<Permissions>) -> bool {
-    permissions
-        .map(|value| value.contains(Permissions::ADMINISTRATOR))
-        .unwrap_or(false)
-}
-
-async fn respond(
-    ctx: &Context,
-    command: &CommandInteraction,
-    ephemeral: bool,
-    content: impl Into<String>,
-) -> Result<(), AppError> {
-    command
-        .create_response(
-            &ctx.http,
-            CreateInteractionResponse::Message(
-                CreateInteractionResponseMessage::new()
-                    .content(content.into())
-                    .ephemeral(ephemeral),
-            ),
-        )
-        .await
-        .map_err(|e| AppError::Internal(e.to_string()))
-}
-
-async fn respond_component(
-    ctx: &Context,
-    component: &ComponentInteraction,
-    ephemeral: bool,
-    content: impl Into<String>,
-) -> Result<(), AppError> {
-    component
-        .create_response(
-            &ctx.http,
-            CreateInteractionResponse::Message(
-                CreateInteractionResponseMessage::new()
-                    .content(content.into())
-                    .ephemeral(ephemeral),
-            ),
-        )
-        .await
-        .map_err(|e| AppError::Internal(e.to_string()))
-}
-
-async fn respond_embed(
-    ctx: &Context,
-    command: &CommandInteraction,
-    ephemeral: bool,
-    embed: CreateEmbed,
-) -> Result<(), AppError> {
-    command
-        .create_response(
-            &ctx.http,
-            CreateInteractionResponse::Message(
-                CreateInteractionResponseMessage::new()
-                    .add_embed(embed)
-                    .ephemeral(ephemeral),
-            ),
-        )
-        .await
-        .map_err(|e| AppError::Internal(e.to_string()))
-}
-
-async fn respond_component_embed(
-    ctx: &Context,
-    component: &ComponentInteraction,
-    ephemeral: bool,
-    embed: CreateEmbed,
-) -> Result<(), AppError> {
-    component
-        .create_response(
-            &ctx.http,
-            CreateInteractionResponse::Message(
-                CreateInteractionResponseMessage::new()
-                    .add_embed(embed)
-                    .ephemeral(ephemeral),
-            ),
-        )
-        .await
-        .map_err(|e| AppError::Internal(e.to_string()))
-}
-
-fn chunk_buttons(buttons: Vec<CreateButton>) -> Vec<Vec<CreateButton>> {
-    let mut rows = Vec::new();
-    let mut current = Vec::new();
-
-    for button in buttons {
-        current.push(button);
-        if current.len() == 5 {
-            rows.push(current);
-            current = Vec::new();
-        }
-    }
-
-    if !current.is_empty() {
-        rows.push(current);
-    }
-
-    rows
-}
-
-fn ping_command() -> CreateCommand {
-    CreateCommand::new("ping")
-        .description("Show the bot gateway latency")
-        .dm_permission(false)
-}
-
-fn register_success_message(
-    base_url: &str,
-    username: &str,
-    api_token: &str,
-    already_registered: bool,
-) -> String {
-    let status_line = if already_registered {
-        format!("You are already registered as `{username}`.")
-    } else {
-        format!("Registered successfully. Username: `{username}`")
-    };
-
-    format!(
-        r#"{status_line}
-API token: `{api_token}`
-
-Get started:
-1. Copy the JSON below.
-2. Replace `YOUR_API_TOKEN` with your token.
-3. In ShareX, go to `Custom uploader settings...` -> `Import` -> `From clipboard`.
-4. Set this uploader as both your **Image uploader** and your **File uploader** in `Destinations`.
-
-```json
-{{
-  "Version": "19.0.0",
-  "Name": "VoidChan",
-  "DestinationType": "ImageUploader, FileUploader, TextUploader",
-  "RequestMethod": "POST",
-  "RequestURL": "{}/api/providers/sharex",
-  "Body": "MultipartFormData",
-  "FileFormName": "file",
-  "Headers": {{
-    "Authorization": "YOUR_API_TOKEN"
-  }},
-  "URL": "{{json:url}}",
-  "ErrorMessage": "{{json:message}}"
-}}
-```"#,
-        base_url.trim_end_matches('/'),
-    )
-}
-
-fn register_command() -> CreateCommand {
-    CreateCommand::new("register")
-        .description("Register your Discord account and get setup instructions")
-        .dm_permission(false)
-}
-
-fn config_command() -> CreateCommand {
-    CreateCommand::new("config")
-        .description("Update your CDN preferences")
-        .add_option(
-            CreateCommandOption::new(
-                CommandOptionType::String,
-                "url_mode",
-                "Preferred copied URL mode",
-            )
-            .required(false)
-            .add_string_choice("View URL (/v)", "v")
-            .add_string_choice("Raw URL (/u)", "u"),
-        )
-        .add_option(
-            CreateCommandOption::new(
-                CommandOptionType::String,
-                "hex_colour",
-                "Preferred embed/meta hex colour like #7289da",
-            )
-            .required(false),
-        )
-        .dm_permission(false)
-}
-
-fn token_command() -> CreateCommand {
-    CreateCommand::new("token")
-        .description("View or regenerate your API token")
-        .add_option(
-            CreateCommandOption::new(
-                CommandOptionType::String,
-                "action",
-                "Choose whether to view or regenerate",
-            )
-            .required(true)
-            .add_string_choice("View", "view")
-            .add_string_choice("Regenerate", "regenerate"),
-        )
-        .dm_permission(false)
-}
-
-fn profile_command() -> CreateCommand {
-    CreateCommand::new("profile")
-        .description("View your CDN profile")
-        .dm_permission(false)
-}
-
-fn files_command() -> CreateCommand {
-    CreateCommand::new("files")
-        .description("List your uploaded files")
-        .dm_permission(false)
-}
-
-fn kv_command() -> CreateCommand {
-    CreateCommand::new("kv")
-        .description("Manage your presence KV entries")
-        .add_option(
-            CreateCommandOption::new(
-                CommandOptionType::SubCommand,
-                "get",
-                "Get a KV value or list all keys",
-            )
-            .add_sub_option(
-                CreateCommandOption::new(CommandOptionType::String, "key", "KV key")
-                    .required(false),
-            ),
-        )
-        .add_option(
-            CreateCommandOption::new(CommandOptionType::SubCommand, "set", "Set a KV value")
-                .add_sub_option(
-                    CreateCommandOption::new(CommandOptionType::String, "key", "KV key")
-                        .required(true),
-                )
-                .add_sub_option(
-                    CreateCommandOption::new(CommandOptionType::String, "value", "KV value")
-                        .required(true),
-                ),
-        )
-        .add_option(
-            CreateCommandOption::new(CommandOptionType::SubCommand, "delete", "Delete a KV value")
-                .add_sub_option(
-                    CreateCommandOption::new(CommandOptionType::String, "key", "KV key")
-                        .required(true),
-                ),
-        )
-        .add_option(CreateCommandOption::new(
-            CommandOptionType::SubCommand,
-            "clear",
-            "Delete all KV entries",
-        ))
-        .add_option(CreateCommandOption::new(
-            CommandOptionType::SubCommand,
-            "export",
-            "Export all KV entries as JSON",
-        ))
-        .dm_permission(false)
-}
-
-fn admin_users_command() -> CreateCommand {
-    CreateCommand::new("admin_users")
-        .description("Admin: list users")
-        .default_member_permissions(Permissions::ADMINISTRATOR)
-        .dm_permission(false)
-}
-
-fn admin_files_command() -> CreateCommand {
-    CreateCommand::new("admin_files")
-        .description("Admin: list uploaded files")
-        .default_member_permissions(Permissions::ADMINISTRATOR)
-        .dm_permission(false)
-}
-
-fn blacklist_command() -> CreateCommand {
-    CreateCommand::new("blacklist")
-        .description("Admin: blacklist or unblacklist a user")
-        .add_option(
-            CreateCommandOption::new(
-                CommandOptionType::String,
-                "discord_user_id",
-                "Discord user ID",
-            )
-            .required(true),
-        )
-        .add_option(
-            CreateCommandOption::new(
-                CommandOptionType::Boolean,
-                "blacklisted",
-                "Whether the user should be blacklisted",
-            )
-            .required(true),
-        )
-        .default_member_permissions(Permissions::ADMINISTRATOR)
-        .dm_permission(false)
-}
-
-fn delete_file_command() -> CreateCommand {
-    CreateCommand::new("delete_file")
-        .description("Admin: delete an uploaded file")
-        .add_option(
-            CreateCommandOption::new(CommandOptionType::String, "file_id", "File ID to delete")
-                .required(true),
-        )
-        .default_member_permissions(Permissions::ADMINISTRATOR)
-        .dm_permission(false)
-}
-
-fn update_disabled(button: CreateButton, disabled: bool) -> CreateButton {
-    button.disabled(disabled)
-}
-
-async fn update_component(
-    ctx: &Context,
-    component: &ComponentInteraction,
-    embed: CreateEmbed,
-    components: Vec<CreateActionRow>,
-) -> Result<(), AppError> {
-    component
-        .create_response(
-            &ctx.http,
-            CreateInteractionResponse::UpdateMessage(
-                CreateInteractionResponseMessage::new()
-                    .embeds(vec![embed])
-                    .components(components),
-            ),
-        )
-        .await
-        .map_err(|e| AppError::Internal(e.to_string()))
-}
-
-fn parse_owned_page_custom_id(custom_id: &str) -> Option<(&str, String, u64)> {
-    let mut parts = custom_id.split(':');
-    let prefix = parts.next()?;
-    let kind = parts.next()?;
-    let owner_id = parts.next()?.to_string();
-    let page = parts.next()?.parse::<u64>().ok()?;
-    if prefix != "page" || parts.next().is_some() {
-        return None;
-    }
-    Some((kind, owner_id, page))
-}
-
-fn clamp_page(page: u64, total_pages: u64) -> u64 {
-    if total_pages == 0 {
-        0
-    } else {
-        page.min(total_pages.saturating_sub(1))
-    }
-}
-
-fn total_pages(total_items: i64, page_size: u64) -> u64 {
-    if total_items <= 0 {
-        0
-    } else {
-        ((total_items as u64) + page_size - 1) / page_size
-    }
-}
-
-fn owned_pagination_components(
-    kind: &str,
-    user_id: u64,
-    current_page: u64,
-    total_pages: u64,
-) -> Vec<CreateActionRow> {
-    if total_pages <= 1 {
-        return Vec::new();
-    }
-
-    vec![CreateActionRow::Buttons(vec![
-        update_disabled(
-            CreateButton::new(format!(
-                "page:{kind}:{user_id}:{}",
-                current_page.saturating_sub(1)
-            ))
-            .label("Previous")
-            .style(ButtonStyle::Secondary),
-            current_page == 0,
-        ),
-        update_disabled(
-            CreateButton::new(format!("page:{kind}:{user_id}:{}", current_page + 1))
-                .label("Next")
-                .style(ButtonStyle::Secondary),
-            current_page + 1 >= total_pages,
-        ),
-    ])]
-}
-
-fn admin_files_pagination_row(current_page: u64, total_pages: u64) -> Option<CreateActionRow> {
-    if total_pages <= 1 {
-        return None;
-    }
-
-    Some(CreateActionRow::Buttons(vec![
-        update_disabled(
-            CreateButton::new(format!(
-                "page:admin_files:{}",
-                current_page.saturating_sub(1)
-            ))
-            .label("Previous")
-            .style(ButtonStyle::Secondary),
-            current_page == 0,
-        ),
-        update_disabled(
-            CreateButton::new(format!("page:admin_files:{}", current_page + 1))
-                .label("Next")
-                .style(ButtonStyle::Secondary),
-            current_page + 1 >= total_pages,
-        ),
-    ]))
-}
