+++ src/lib.rs
@@ -1,0 +1,6 @@
+pub mod app;
+pub mod domain;
+pub mod http;
+pub mod integrations;
+pub mod services;
+pub mod utils;
