+++ src/services/presence.rs
